requirements.txt
kafka-python>=2.0.2
pyspark>=3.5.0
influxdb-client>=1.36.0
pandas>=2.0.3
pyyaml>=6.0